Make pages full screen height
Ensure that pages always take up full screen height, to avoid
whitespace at bottom of the page (especially problematic in
dark theme, where the whitespace is actually white, and
clashes with the theme color).

Opted to solve this by making the page layout wrapper have
full screen height, and adding the extra space to the bottom
of the page, so the whitespace at the bottom is themed correctly.

Decided not to use grid to stretch the main page content to
fill the empty space, since the main content is fairly complex
and results in large empty areas in the middle on small pages
(like the 404 page).

Also removing the rogue wrapper div in ClientOnly.

Resolves #42